tests wip

+++ service/src/main/java/bio/terra/pipelines/app/configuration/external/GcsConfiguration.java
@@ -3,5 +3,4 @@ package bio.terra.pipelines.app.configuration.external;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 @ConfigurationProperties(prefix = "gcs")
-public record GcsConfiguration(
-    Integer signedUrlPutDurationHours, Integer signedUrlGetDurationHours) {}
+public record GcsConfiguration(long signedUrlPutDurationHours, long signedUrlGetDurationHours) {}
